+++ src/components/index.js
@@ -1,11 +1,13 @@
 import EditOrgs from './EditOrgs'
 import AddOrg from './AddOrg'
 import OrgsList from './OrgsList'
+import LinksList from './LinksList'
 import OrgsListEditable from './OrgsListEditable'
 import { Btn, Bttn } from './Btn'
 import OrgAction from './OrgAction'
 import OrgActionIdModal from './OrgActionIdModal'
 import OrgsListItem from './OrgsListItem'
+import LinksListItem from './LinksListItem'
 import OrgsListItemEditable from './OrgsListItemEditable'
 import OrgColorAction from './OrgColorAction'
 import OrgEditAction from './OrgEditAction'
@@ -13,6 +15,7 @@ import SortableList from './SortableList'
 import SortableItem from './SortableItem'
 import Head from './Head'
 import HeadLine from './HeadLine'
+import HeadLink from './HeadLink'
 import IconAction from './IconAction'
 import Options from './Options'
 
@@ -20,18 +23,21 @@ export {
   EditOrgs,
   AddOrg,
   OrgsList,
+  LinksList,
   OrgsListEditable,
   Btn,
   Bttn,
   OrgAction,
   OrgActionIdModal,
   OrgsListItem,
+  LinksListItem,
   OrgsListItemEditable,
   OrgColorAction,
   OrgEditAction,
   SortableList,
   SortableItem,
   Head,
+  HeadLink,
   HeadLine,
   IconAction,
   Options,