+++ src/pages/Popup/Popup.jsx
@@ -1,25 +1,38 @@
-import React from 'react'
+import React, { useState } from 'react'
 import { useChromeStorageLocal } from 'use-chrome-storage'
-import { VStack, Box, Alert, AlertIcon } from '@chakra-ui/react'
-import { OrgsList, Head, HeadLine, Options } from '../../components'
+import { VStack, Box, Flex } from '@chakra-ui/react'
+import { OrgsList, LinksList, Head, HeadLink, Options } from '../../components'
+import { defaultLinks } from '../../utils/defaultLinks.js'
 
 const Popup = () => {
-  const [orgs, setOrgs, isPersistent, error] = useChromeStorageLocal('SalesforceOrgs', [])
+  const [orgs, setOrgs] = useChromeStorageLocal('SalesforceOrgs', [])
+  const [links, setLinks] = useChromeStorageLocal('SalesforceLinks', [])
+  const [isOpen, setOpen] = useState(true)
 
   return (
     <VStack p={0} w='440px' minH='230px'>
       <Head>
-        <HeadLine text='Salesforce Orgs' />
+        <Flex w='100%'>
+          <HeadLink
+            fontWeight={isOpen ? 'bold' : 'normal'}
+            action={() => {
+              setOpen(true)
+            }}
+            text='Salesforce Orgs'
+          />
+          <HeadLink
+            mx={5}
+            fontWeight={isOpen ? 'normal' : 'bold'}
+            action={() => {
+              setOpen(false)
+            }}
+            text='Links'
+          />
+        </Flex>
         <Options orgs={orgs} setOrgs={setOrgs} />
       </Head>
       <Box w='100%' p={4}>
-        <OrgsList items={orgs} />
-        {!isPersistent && (
-          <Alert status='error'>
-            <AlertIcon />
-            Error writing to the chrome.storage: {error}
-          </Alert>
-        )}
+        {isOpen ? <OrgsList items={orgs} /> : <LinksList defaultLinks={defaultLinks} links={links} />}
       </Box>
     </VStack>
   )