+++ src/utils/defaultLinks.js
@@ -1,0 +1,26 @@
+const defaultLinks = {
+    'links': [
+        {
+            'id': 'link-1',
+            'name': 'Hands-On Orgs',
+            'href': 'trailhead.salesforce.com/users/profiles/orgs'
+        },
+        {
+            'id': 'link-2',
+            'name': 'Salesforce Help',
+            'href': 'help.salesforce.com'
+        },
+        {
+            'id': 'link-3',
+            'name': 'Known Issues',
+            'href': 'issues.salesforce.com'
+        },
+        {
+            'id': 'link-4',
+            'name': 'Salesforce Status',
+            'href': 'status.salesforce.com'
+        }
+    ]
+}
+
+module.exports = { defaultLinks }